# specs/001-telco-adb-automation/contracts/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Telco ADB Automation API",
    "description": "REST API for managing devices, flows, and executions in the Telco ADB Automation Desktop Framework",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/executions": {
      "get": {
        "summary": "List executions",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["pending", "running", "completed", "failed", "cancelled"] }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "List of executions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Execution" }
                }
              }
            }
          }
        }
      }
    },
    "/executions/{execution_id}": {
      "get": {
        "summary": "Get execution details",
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Execution details with steps",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExecutionDetail" }
              }
            }
          },
          "404": {
            "description": "Execution not found"
          }
        }
      }
    },
...
    "/reports/{execution_id}": {
      "get": {
        "summary": "Download execution report",
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "format",
            "in": "query",
            "schema": { "type": "string", "enum": ["pdf", "csv", "html"], "default": "pdf" }
          }
        ],
        "responses": {
          "200": {
            "description": "Report file",
            "content": {
              "application/pdf": {},
              "text/csv": {},
              "text/html": {}
            }
          },
          "404": {
            "description": "Report not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Execution": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "flow_id": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["pending", "running", "completed", "failed", "cancelled"] },
          "start_time": { "type": "string", "format": "date-time", "nullable": true },
          "end_time": { "type": "string", "format": "date-time", "nullable": true },
          "progress_percentage": { "type": "integer" },
          "current_step": { "type": "string", "nullable": true },
          "devices": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ExecutionDevice" }
          }
        }
      },
      "ExecutionDevice": {
        "type": "object",
        "properties": {
          "device_id": { "type": "string" },
          "status": { "type": "string", "enum": ["assigned", "running", "completed", "failed"] },
          "start_time": { "type": "string", "format": "date-time", "nullable": true },
          "end_time": { "type": "string", "format": "date-time", "nullable": true },
          "error_message": { "type": "string", "nullable": true }
        }
      },
      "ExecutionDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Execution" },
          {
            "type": "object",
            "properties": {
              "steps": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ExecutionStep" }
              }
            }
          }
        ]
      },
      "ExecutionStep": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "module_id": { "type": "string" },
          "device_id": { "type": "string" },
          "step_index": { "type": "integer" },
          "status": { "type": "string", "enum": ["pending", "running", "completed", "failed", "skipped"] },
          "start_time": { "type": "string", "format": "date-time", "nullable": true },
          "end_time": { "type": "string", "format": "date-time", "nullable": true },
          "output_data": { "type": "object", "nullable": true },
          "error_code": { "type": "string", "nullable": true },
          "screenshot_path": { "type": "string", "nullable": true }
        }
      }
    }
  }
}